Fix home controller
Sửa lại các controller của trang home
Tách các class DAO để quản lý source code dễ hơn
