+++ .gitignore
@@ -1,2 +1,3 @@
 /nbproject/private/
 /build/
+/dist/
