+++ nbproject/private/private.properties
@@ -1,10 +1,10 @@
-deploy.ant.properties.file=C:\\Users\\Dell\\AppData\\Roaming\\NetBeans\\8.2rc\\config\\GlassFishEE6\\Properties\\gfv3-380332873.properties
+deploy.ant.properties.file=C:\\Users\\Hfyl\\AppData\\Roaming\\NetBeans\\8.2rc\\tomcat90.properties
 j2ee.platform.is.jsr109=true
-j2ee.server.domain=C:/Users/Dell/AppData/Roaming/NetBeans/8.2rc/config/GF_4.1.1/domain1
-j2ee.server.home=C:/Program Files/glassfish-4.1.1/glassfish
-j2ee.server.instance=[C:\\Program Files\\glassfish-4.1.1\\glassfish;C:\\Program Files\\glassfish-4.1.1\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
+j2ee.server.domain=C:/Users/Hfyl/AppData/Roaming/NetBeans/8.2rc/config/GF_4.1.1/domain1
+j2ee.server.home=C:/apache-tomcat-9.0.45
+j2ee.server.instance=tomcat90:home=C:\\apache-tomcat-9.0.45
 j2ee.server.middleware=C:/Program Files/glassfish-4.1.1
 javac.debug=true
 javadoc.preview=true
-selected.browser=SL[/Browsers/MicrosoftEdgeBrowser
-user.properties.file=C:\\Users\\Dell\\AppData\\Roaming\\NetBeans\\8.2rc\\build.properties
+selected.browser=Chrome
+user.properties.file=C:\\Users\\Hfyl\\AppData\\Roaming\\NetBeans\\8.2rc\\build.properties
