+++ src/java/controller/util/BooksDAO.java
@@ -3,10 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package dao;
+package controller.util;
 
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -14,40 +13,34 @@ import java.util.ArrayList;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import model.Book;
+import model.DBConnection;
 
 /**
  *
- * @author admin
+ * @author Hfyl
  */
-public class DBContext {
-    Connection connection;
-    
-    public DBContext() {
+public class BooksDAO {
 
-        try {
-            String user = "sa";
-            String pass = "Anhvu123";
-            String url = "jdbc:sqlserver://localhost:1433;databaseName=BookShops";
-            Class.forName("com.microsoft.sqlserver.jdbc.SQLServerDriver");
-            connection = DriverManager.getConnection(url, user, pass);
-        } catch (SQLException ex) {
-            Logger.getLogger(DBContext.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (ClassNotFoundException ex) {
-            Logger.getLogger(DBContext.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
+    private Connection con;
+    private PreparedStatement ps;
+    private ResultSet rs;
+    private AuthorsDAO au;
+
     public ArrayList<Book> getAllBooks() {
         ArrayList<Book> books = new ArrayList<>();
         String sql = "SELECT *\n "
                 + "FROM\n "
                 + "Products\n ";
-
-        PreparedStatement st;
+        
         try {
-            st = connection.prepareStatement(sql);
+            con = DBConnection.open();
+            ps = con.prepareStatement(sql);
+            rs = ps.executeQuery();
 
-            ResultSet rs = st.executeQuery();
+            //initialize AuthorsDAO object
+            au = new AuthorsDAO();
+
+            //assign data to books
             while (rs.next()) {
                 Book book = new Book();
                 book.setProductID(rs.getInt("ProductID"));
@@ -56,44 +49,18 @@ public class DBContext {
                 book.setDescription(rs.getString("Description"));
                 book.setUnitPrice(rs.getFloat("UnitPrice"));
 
-                book.setAuthors(getAuthorsByBookId(book.getProductID()));
+                // get book authors via AuthorsDAO
+                book.setAuthors(au.getAuthorsByBookId(book.getProductID()));
                 books.add(book);
             }
 
         } catch (SQLException ex) {
-            Logger.getLogger(DBContext.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(BooksDAO.class.getName()).log(Level.SEVERE, null, ex);
+        } finally {
+            DBConnection.close(con, ps, rs);
         }
 
         return books;
-    }
-
-    public ArrayList<String> getAuthorsByBookId(int bookId) {
-        ArrayList<String> authors = new ArrayList();
-
-        String sql = "SELECT *\n"
-                + "from ProAu,Authors\n"
-                + " where ProAu.AuthorID = Authors.AuthorID\n"
-                + " and ProAu.ProductID = ?";
-
-        PreparedStatement st;
-
-        try {
-            st = connection.prepareStatement(sql);
-
-            st.setInt(1, bookId);
-
-            ResultSet rs = st.executeQuery();
-
-            while (rs.next()) {
-                String author = rs.getString("AuthorName");
-
-                authors.add(author);
-            }
-        } catch (SQLException ex) {
-            Logger.getLogger(DBContext.class.getName()).log(Level.SEVERE, null, ex);
-        }
-
-        return authors;
     }
 
     public ArrayList<Book> getBestSellerBooks() {
@@ -102,11 +69,15 @@ public class DBContext {
                 + "from Products\n"
                 + "order by UnitInStock desc";
 
-        PreparedStatement st;
         try {
-            st = connection.prepareStatement(sql);
+            con = DBConnection.open();
+            ps = con.prepareStatement(sql);
+            rs = ps.executeQuery();
 
-            ResultSet rs = st.executeQuery();
+            //initialize AuthorsDAO object
+            au = new AuthorsDAO();
+
+            //assign data to books
             while (rs.next()) {
                 Book book = new Book();
                 book.setProductID(rs.getInt("ProductID"));
@@ -114,12 +85,16 @@ public class DBContext {
                 book.setPathImage(rs.getString("ImagePath"));
                 book.setDescription(rs.getString("Description"));
                 book.setUnitPrice(rs.getFloat("UnitPrice"));
-                book.setAuthors(getAuthorsByBookId(book.getProductID()));
+
+                // get book authors via AuthorsDAO
+                book.setAuthors(au.getAuthorsByBookId(book.getProductID()));
                 books.add(book);
             }
 
         } catch (SQLException ex) {
-            Logger.getLogger(DBContext.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(BooksDAO.class.getName()).log(Level.SEVERE, null, ex);
+        } finally {
+            DBConnection.close(con, ps, rs);
         }
 
         return books;
@@ -132,11 +107,15 @@ public class DBContext {
                 + "where Products.CategoryID=Category.CategoryID\n"
                 + "and Products.CategoryID=1";
 
-        PreparedStatement st;
         try {
-            st = connection.prepareStatement(sql);
-            
-            ResultSet rs = st.executeQuery();
+            con = DBConnection.open();
+            ps = con.prepareStatement(sql);
+            rs = ps.executeQuery();
+
+            //initialize AuthorsDAO object
+            au = new AuthorsDAO();
+
+            //assign data to books
             while (rs.next()) {
                 Book book = new Book();
                 book.setProductID(rs.getInt("ProductID"));
@@ -144,17 +123,21 @@ public class DBContext {
                 book.setPathImage(rs.getString("ImagePath"));
                 book.setDescription(rs.getString("Description"));
                 book.setUnitPrice(rs.getFloat("UnitPrice"));
-                book.setAuthors(getAuthorsByBookId(book.getProductID()));
+
+                // get book authors via AuthorsDAO
+                book.setAuthors(au.getAuthorsByBookId(book.getProductID()));
                 books.add(book);
             }
 
         } catch (SQLException ex) {
-            Logger.getLogger(DBContext.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(BooksDAO.class.getName()).log(Level.SEVERE, null, ex);
+        } finally {
+            DBConnection.close(con, ps, rs);
         }
 
         return books;
     }
-    
+
     public ArrayList<Book> getBookByBook() {
         ArrayList<Book> books = new ArrayList<>();
         String sql = "select * \n"
@@ -162,10 +145,15 @@ public class DBContext {
                 + "where Products.CategoryID=Category.CategoryID\n"
                 + "and Products.CategoryID=2";
 
-        PreparedStatement st;
         try {
-            st = connection.prepareStatement(sql);
-            ResultSet rs = st.executeQuery();
+            con = DBConnection.open();
+            ps = con.prepareStatement(sql);
+            rs = ps.executeQuery();
+
+            //initialize AuthorsDAO object
+            au = new AuthorsDAO();
+
+            //assign data to books
             while (rs.next()) {
                 Book book = new Book();
                 book.setProductID(rs.getInt("ProductID"));
@@ -173,17 +161,21 @@ public class DBContext {
                 book.setPathImage(rs.getString("ImagePath"));
                 book.setDescription(rs.getString("Description"));
                 book.setUnitPrice(rs.getFloat("UnitPrice"));
-                book.setAuthors(getAuthorsByBookId(book.getProductID()));
+
+                // get book authors via AuthorsDAO
+                book.setAuthors(au.getAuthorsByBookId(book.getProductID()));
                 books.add(book);
             }
 
         } catch (SQLException ex) {
-            Logger.getLogger(DBContext.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(BooksDAO.class.getName()).log(Level.SEVERE, null, ex);
+        } finally {
+            DBConnection.close(con, ps, rs);
         }
 
         return books;
     }
-    
+
     public ArrayList<Book> getBookByNovel() {
         ArrayList<Book> books = new ArrayList<>();
         String sql = "select * \n"
@@ -191,10 +183,15 @@ public class DBContext {
                 + "where Products.CategoryID=Category.CategoryID\n"
                 + "and Products.CategoryID=3";
 
-        PreparedStatement st;
         try {
-            st = connection.prepareStatement(sql);
-            ResultSet rs = st.executeQuery();
+            con = DBConnection.open();
+            ps = con.prepareStatement(sql);
+            rs = ps.executeQuery();
+
+            //initialize AuthorsDAO object
+            au = new AuthorsDAO();
+
+            //assign data to books
             while (rs.next()) {
                 Book book = new Book();
                 book.setProductID(rs.getInt("ProductID"));
@@ -202,14 +199,19 @@ public class DBContext {
                 book.setPathImage(rs.getString("ImagePath"));
                 book.setDescription(rs.getString("Description"));
                 book.setUnitPrice(rs.getFloat("UnitPrice"));
-                book.setAuthors(getAuthorsByBookId(book.getProductID()));
+
+                // get book authors via AuthorsDAO
+                book.setAuthors(au.getAuthorsByBookId(book.getProductID()));
                 books.add(book);
             }
 
         } catch (SQLException ex) {
-            Logger.getLogger(DBContext.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(BooksDAO.class.getName()).log(Level.SEVERE, null, ex);
+        } finally {
+            DBConnection.close(con, ps, rs);
         }
 
         return books;
     }
+
 }
