+++ web/META-INF/context.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<Context path="/BooksShopOnline"/>
+<Context path="/Swp391-SE1515-NET-Group4"/>
