--- web/index.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<!--
-To change this license header, choose License Headers in Project Properties.
-To change this template file, choose Tools | Templates
-and open the template in the editor.
--->
-<html>
-    <head>
-        <title>TODO supply a title</title>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <link href="public/css/index.css" rel="stylesheet" type="text/css"/>
-    </head>
-    <body>
-        <div>
-            <h3 class="topic">
-                Swp391-SE1515-NET-Group4
-            </h3>
-            <ul class="topic-note">
-                <li>
-                    - Tạo Pull request khi push
-                </li>
-                <li>
-                    - Check Branch mình làm
-                    <br>
-                    - format đặt tên branch feature/"name function"
-                    <br>
-                    VD: feature/LandingPage
-                </li>
-            </ul>
-        </div>
-    </body>
-</html>
